TDD - Question value appropriateness
Controller can determine whether a value is less than 0 or greater than
5 by returning an illegalArgumentException when either condition is met
